PF tuning

+++ Filters/Fusion/Lib/system.h
@@ -9,8 +9,7 @@
 #define FLIGHTCOMPUTER_SYSTEM_H
 
 #include <stdbool.h>
-#include <glob.h>
-#include "config.h"
+#include <stdlib.h>
 
 typedef struct {
     float roll_angle, roll_deriv; // Orientation and speed of the plane around the roll (forward) axis (degree)
@@ -19,7 +18,7 @@ typedef struct {
     float speed; // The absolute speed of the plane
     float altitude; // The absolute altitude of the plane
     float altitude_above_ground; // The altitude above ground
-    float lat, lon; // The global position of the plane
+    long double lat, lon; // The global position of the plane
 } system_state_t;
 
 typedef struct {
@@ -37,7 +36,7 @@ typedef struct {
     float altitude_baro; // The absolute altitude, as measured by the barometer
     float altitude_gps; // The absolute altitude, as measured by the gps
     float distance_ground; // The distance to the ground, as measured by the distance sensor
-    float lat, lon; // The position, acquired by the GPS
+    long double lat, lon; // The position, acquired by the GPS
 } measurement_t;
 
 typedef struct {
